--- src/development/vortex/development/exporter/utils/onnx/graph_ops/base_ops.py
@@ -1,1 +1,0 @@
-from vortex.runtime.onnx.graph_ops.base_ops import GraphOpsBase
