+++ src/runtime/vortex/runtime/onnx/graph_ops/embed_output_format.py
@@ -7,17 +7,6 @@ from typing import Union, Dict, List
 
 logger = logging.getLogger(__name__)
 
-def embed_output_format(model : onnx.ModelProto, output_format : Dict[str,Union[List[int],int]]) -> onnx.ModelProto:
-    """
-    embed output_format to model as `Constants`
-    """
-    logger.info("embedding output_format to model")
-    output_format_constants, output_format_value_info = make_output_format(output_format)
-    for output_format_constant, value_info in zip(output_format_constants, output_format_value_info):
-        model.graph.node.append(output_format_constant)
-        model.graph.output.append(value_info)
-    return model
-
 def get_ops(*args, **kwargs) :
     return embed_output_format
 
@@ -33,6 +22,18 @@ class EmbedOutputFormat(GraphOpsBase):
         """        
         self.output_format = output_format
     
+    @classmethod
+    def embed_output_format(cls, model : onnx.ModelProto, output_format : Dict[str,Union[List[int],int]]) -> onnx.ModelProto:
+        """
+        embed output_format to model as `Constants`
+        """
+        logger.info("embedding output_format to model")
+        output_format_constants, output_format_value_info = make_output_format(output_format)
+        for output_format_constant, value_info in zip(output_format_constants, output_format_value_info):
+            model.graph.node.append(output_format_constant)
+            model.graph.output.append(value_info)
+        return model
+    
     def run(self, model: onnx.ModelProto) -> onnx.ModelProto:
         """Actually embed output_format to model
 
@@ -42,4 +43,4 @@ class EmbedOutputFormat(GraphOpsBase):
         Returns:
             onnx.ModelProto: transformed model
         """        
-        return embed_output_format(model, **vars(self))
+        return self.embed_output_format(model, **vars(self))
